requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
feedparser==5.2.1
future==0.17.1
idna==2.8
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
soupsieve==1.8
SQLAlchemy==1.2.18
urllib3==1.24.1
